change acceptance check, no longer a test
