+++ cd4ml/accept_model.py
@@ -1,0 +1,32 @@
+import os
+
+import mlflow
+import mlflow.tracking
+
+
+def get_latest_executed_run(df_of_runs):
+    filtered_dataframe = df_of_runs[df_of_runs["tags.mlflow.runName"] == os.environ["BUILD_NUMBER"]]
+    assert len(filtered_dataframe) == 1
+    return filtered_dataframe
+
+
+def get_metric(metric_name, df_of_single_run):
+    print(df_of_single_run)
+    return df_of_single_run["metrics.%s" % metric_name][0]
+
+
+def check_model_performance(metric_name, threshold):
+    mlflow.set_tracking_uri(os.environ["MLFLOW_TRACKING_URL"])
+    experiment = mlflow.get_experiment_by_name("local")
+    runs = mlflow.search_runs(experiment_ids=experiment.experiment_id)
+    last_run_record = get_latest_executed_run(runs)
+    metric_value = get_metric(metric_name, last_run_record)
+    run_name = last_run_record["tags.mlflow.runName"][0]
+    template = "{metric_name} for Run '{run_name}' was too low for model acceptance," \
+               " {metric_name}: threshold: {threshold}"
+    message = template.format(matric_name=metric_name,
+                              run_name=run_name,
+                              metric_value=metric_value,
+                              threshold=threshold)
+
+    assert metric_value > threshold, message
